+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - iOS users should now open the adresses with maps.app !2236 @sefan 
 - Force content which is too broad to scroll, so the the Topbar will never leave the screen width !2258 @NINI1988
 - A wall post with multiple lines will use the complete available size to display its text !2277 @NINI1988
+- Restrictions on the view of the store list in profile !2273 @chriswalg
 
 ## Refactoring
 - Currently running events are shown within the event overview until their end !2162 @merhoc